requirements.txt
requests
beautifulsoup4
feedparser
python-dateutil
pyyaml
rapidfuzz
python-dotenv
